DB-11826 address comments.
