+++ splice_machine/src/main/java/com/splicemachine/derby/impl/sql/execute/operations/ScanOperation.java
@@ -134,9 +134,9 @@ public abstract class ScanOperation extends SpliceBaseOperation {
                 defaultValueMapItem,
                 numUnusedLeadingIndexFields
         );
-        if(pastTxn != -2) {
+        if(pastTxn != TransactionController.TIME_TRAVEL_UNSET) {
             this.pastTx = pastTxn;
-            if(this.pastTx == -1) {
+            if(this.pastTx == TransactionController.TIME_TRAVEL_OLDEST) {
                 this.pastTx = SIConstants.OLDEST_TIME_TRAVEL_TX; // force going back to the oldest transaction instead of ignoring it.
             }
         } else {
